% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{calc_limit}
\alias{calc_limit}
\title{Internally calculates a limit based on given data and perhaps variance data}
\usage{
calc_limit(data, variance_data = NULL, limit = 3.3, .corrected = FALSE)
}
\arguments{
\item{data}{The data to use}

\item{variance_data}{The variance data to use. If NULL, the data is used instead}

\item{limit}{The limit to calculate (e.g. 3.3 for LOD, 10 for LOQ)}

\item{.corrected}{Whether to use the unbiased standard deviation or not}
}
\value{
The calculated limit
}
\description{
Internally calculates a limit based on given data and perhaps variance data
}
\examples{
calc_limit(rnorm(10, 0, 4))
calc_limit(rnorm(10, 0, 4), .corrected = TRUE)
calc_limit(rnorm(10, 0, 4), rnorm(10, 0, 4))
}
