% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.derivatives.R
\name{dataprep.derivatives}
\alias{dataprep.derivatives}
\title{Data Preparation for `trendplot.gam`}
\usage{
dataprep.derivatives(object, data, fitted = TRUE, ...)
}
\arguments{
\item{object}{A `mgcv::gam` or `mgcv::gamm` object.}

\item{data}{Data frame source of the \code{object} argument.}

\item{fitted}{A logical argument. \code{TRUE} dictates that derivatives are calculated from equally separated, weekly fitted values \code{(DEFAULT)}. \code{FALSE} dictates that derivatives are calculated from predictor values in \code{data}.}

\item{...}{Additional arguments passed onto `gratia::derivatives`.
However, you cannot change the object, data, select, partial_match, type, and interval arguments as this function have preset options.}
}
\description{
This is used to prepare the data for the \code{`trendplot.gam`} function.
It is a wrapper for various functions in \pkg{tidyverse}, \pkg{mgcv}, and \pkg{gratia}. Please cite those packages.
}
\examples{
data1 <- data.frame(
DATE = c("1988-06-27", "1988-08-25", "1990-05-03", "1990-05-31", "1990-06-26", "1988-06-27", "1988-08-25", "1990-05-03", "1990-05-31", "1990-06-26"),
NUMDATE = c(6752, 6811, 7427, 7455, 7481, 6752, 6811, 7427, 7455, 7481),
JULIAN = c(179, 238, 123, 151, 177, 179, 238, 123, 151, 177),
GROUP = c('A', 'A', 'A', 'A', 'A', 'B', 'B', 'B', 'B', 'B'),
RESPONSE = c(22, 49, 75, 41, 36, 45, 2, 20, 31, 40)
)

mod1 <- mgcv::gam(RESPONSE ~ GROUP + s(NUMDATE, by = GROUP, k = kyears, bs = 'tp') +
s(JULIAN, by = GROUP, k = kmonths, bs = 'cc'),
data = data1, method = 'REML', family = 'nb',
knots = list(JULIAN = c(1,365)))

dataprep.derivatives(mod1, data1, fitted = TRUE)

}
