% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{get_intermed}
\alias{get_intermed}
\title{Get assay intermediate precision CV measure}
\usage{
get_intermed(model)
}
\arguments{
\item{model}{A random-effects model}
}
\value{
The intermediate precision (as percent RSD)
}
\description{
Returns the intermediate precision based on the sum of all the variances of the random intercepts.
Works for any lme4 object, but make sure it's only got a fixed intercepts and no other fixed terms.
Random ones can be anything, as long as they describe the model hierarchy properly.
}
