% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{get_intermed_levels}
\alias{get_intermed_levels}
\title{Get assay intermediate precision factors CV measure}
\usage{
get_intermed_levels(model)
}
\arguments{
\item{model}{A random-effects model}
}
\value{
The components for the intermediate precision (as percent RSD)
}
\description{
Returns the added relative standard deviation based on the variances of the random intercepts in the model.
Works for any lme4 object, but make sure it's only got a fixed intercepts and no other fixed terms.
Random ones can be anything, as long as they describe the model hierarchy properly.
}
