% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCID50.R
\name{get_logTCID}
\alias{get_logTCID}
\title{TCID50 Computing Functions - logTCID50}
\usage{
get_logTCID(x, d = 1, log_dilution)
}
\arguments{
\item{x}{A vector containing proportion of positive samples}

\item{d}{Log-Dilution factor difference, 1 for 1:10 dilutions}

\item{log_dilution}{the log-dilution of the assay}
}
\description{
TCID50 Computing Functions - logTCID50
}
