% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCID50.R
\name{get_logTCID_SE}
\alias{get_logTCID_SE}
\title{TCID50 Computing Functions - logTCID50 Standard Error}
\usage{
get_logTCID_SE(x, d = 1, total)
}
\arguments{
\item{x}{A vector containing proportion of positive samples}

\item{d}{Log-Dilution factor difference, 1 for 1:10 dilutions}

\item{total}{Number of technical replicates (samples for a given dilution)}
}
\description{
TCID50 Computing Functions - logTCID50 Standard Error
}
