% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCID50.R
\name{get_PI_ratio}
\alias{get_PI_ratio}
\title{TCID50 Computing Functions - P:I Ratio}
\usage{
get_PI_ratio(x, d = 1, log_dilution, vol_inoculum, vec_concentration)
}
\arguments{
\item{x}{A vector containing proportion of positive samples}

\item{d}{Log-Dilution factor difference, 1 for 1:10 dilutions}

\item{log_dilution}{A vector of log10-dilutions (-1 for 1/10, -2 for 1/100, etc.)}

\item{vol_inoculum}{Volume Inoculum (in ml)}

\item{vec_concentration}{Vector concentration (GC/ml)}
}
\description{
TCID50 Computing Functions - P:I Ratio
}
