% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{get_repeat}
\alias{get_repeat}
\title{Get assay repeatability CV measure}
\usage{
get_repeat(model)
}
\arguments{
\item{model}{A random-effects model}
}
\value{
The repeatability (as percent RSD)
}
\description{
Returns the repeatability based on the residual variance in the model.
Works for any lme4 object, but make sure it's only got a fixed intercepts and no other fixed terms.
Random ones can be anything as long as they describe the model hierarchy properly.
}
