% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCID50.R
\name{get_TCID_CI}
\alias{get_TCID_CI}
\title{TCID50 Computing Functions - TCID50 Confidence Interval}
\usage{
get_TCID_CI(x, d = 1, log_dilution, vol_inoculum, total, alpha = 0.05)
}
\arguments{
\item{x}{A vector containing proportion of positive samples}

\item{d}{Log-Dilution factor difference, 1 for 1:10 dilutions}

\item{log_dilution}{A vector of log10-dilutions (-1 for 1/10, -2 for 1/100, etc.)}

\item{vol_inoculum}{Volume Inoculum (in ml)}

\item{total}{Number of technical replicates (samples for a given dilution)}

\item{alpha}{Significance threshold (0.05 alpha corresponds to 95\% CI)}
}
\description{
TCID50 Computing Functions - TCID50 Confidence Interval
}
