% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCID50.R
\name{get_TCID_SE}
\alias{get_TCID_SE}
\title{TCID50 Computing Functions - TCID50 Standard Error}
\usage{
get_TCID_SE(x, d = 1, log_dilution, vol_inoculum, total)
}
\arguments{
\item{x}{A vector containing proportion of positive samples}

\item{d}{Log-Dilution factor difference, 1 for 1:10 dilutions}

\item{log_dilution}{A vector of log10-dilutions (-1 for 1/10, -2 for 1/100, etc.)}

\item{vol_inoculum}{Volume Inoculum (in ml)}

\item{total}{Number of technical replicates (samples for a given dilution)}
}
\description{
TCID50 Computing Functions - TCID50 Standard Error
}
