% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insetMap.R
\name{insetMap}
\alias{insetMap}
\title{insetMap}
\usage{
insetMap(map, rast)
}
\arguments{
\item{map}{A ggplot map}

\item{rast}{A raster defining the inset area}
}
\value{
Map with inset
}
\description{
insetMap
}
\examples{
insetMap(map)
}
