% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{kable_style}
\alias{kable_style}
\title{Turn kable output into a pretty kable output}
\usage{
kable_style(x, full = F)
}
\arguments{
\item{x}{An html-styled table, as produced by kable(format = "html")}

\item{full}{Boolean value indicating whether the table should span the entire width of the screen (TRUE) or not (default, FALSE)}
}
\value{
A pretty kable output, to use in markdown reports; requires kableExtra
}
\description{
Turn kable output into a pretty kable output
}
\examples{
\dontrun{
mtcars \%>\% kable() \%>\% kable_style()
}
}
