% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapExtra.R
\name{mapExtra}
\alias{mapExtra}
\title{mapExtra}
\usage{
mapExtra(map, land = TRUE, northArrow = TRUE, scalebar = TRUE)
}
\arguments{
\item{map}{A ggplot-map}

\item{land}{Includes a background map}

\item{northArrow}{Includes a north arrow}

\item{scalebar}{Includes a scale bar}
}
\value{
Same as input
}
\description{
mapExtra
}
\examples{
mapExtra(map, scalebar = FALSE)
}
