% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_in_excel.R
\name{open_in_excel}
\alias{open_in_excel}
\title{Open in Excel}
\usage{
open_in_excel(.data)
}
\arguments{
\item{.data}{A dataframe that will be opened in Excel.}
}
\value{
Same as input
}
\description{
Open in Excel
}
\examples{
open_in_excel(iris)
}
