% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.kendall.R
\name{pairwise.kendall}
\alias{pairwise.kendall}
\title{Pairwise Synchrony Analysis}
\usage{
pairwise.kendall(data)
}
\arguments{
\item{data}{A wide-form \code{data.frame} object where columns are variables and rows are samples.}
}
\description{
This function tests for lake synchrony or can be used for multiple variables using Kendall tau-b's coefficient. \cr\cr

This is primarily a wrapper for the \code{`Kendall`} function in \pkg{Kendall} and various functions in \pkg{tidyverse} for data management. Please cite the listed packages.
}
\examples{
data <- data.frame(DATE = c("1988-06-27", "1988-08-25", "1990-05-03", "1990-05-31", "1990-06-26", "1990-08-21", "1992-07-16", "1992-08-10", "1994-05-25", "1994-08-03"),
var1 = c(22, 49, 75, 41, 36, 45, 2, 20, 31, 40),
var2 = c(70, 58, 96, 28, 78, 79, 53, 66, 84, 80))
pairwise.kendall(data[-1])

}
