% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual_plots.R
\name{plot_accuracy}
\alias{plot_accuracy}
\title{Create accuracy plot}
\usage{
plot_accuracy(data, ref_value = NULL, limits = c(0.8, 1.25), .log = TRUE)
}
\arguments{
\item{data}{Data to plot}

\item{ref_value}{Reference value to compare against}

\item{limits}{Limits to plot}

\item{.log}{Whether to log-transform the y variable (default = TRUE)}
}
\value{
ggplot2 object with accuracy evaluation
}
\description{
Create accuracy plot
}
