% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual_plots.R
\name{plot_comparison}
\alias{plot_comparison}
\title{Create a comparison Bland-Altman plot}
\usage{
plot_comparison(x, y, .log = TRUE)
}
\arguments{
\item{x}{Method 1 data}

\item{y}{Method 2 data}

\item{.log}{Whether to log-transform the y variable and present the result as percent difference (default = TRUE)}
}
\value{
ggplot2 object with Bland-Altman plot
}
\description{
Create a comparison Bland-Altman plot
}
