% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual_plots.R
\name{plot_intermediate_precision}
\alias{plot_intermediate_precision}
\title{Create an intermediate precision plot}
\usage{
plot_intermediate_precision(data, type = "A", limit = 0.3)
}
\arguments{
\item{data}{Data to plot}

\item{type}{Type of intermediate precision to plot, "A" = intermediate precision heatmap, "B" = traditional plot.}

\item{limit}{The acceptance criteria limit}
}
\value{
ggplot2 object with intermediate precision evaluation
}
\description{
Create an intermediate precision plot
}
