% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual_plots.R
\name{plot_limit}
\alias{plot_limit}
\title{Create a LoD/LoQ plot}
\usage{
plot_limit(data, limit = NULL, .setcolour = NULL, .log.y = TRUE)
}
\arguments{
\item{data}{Data to plot}

\item{limit}{Set the limit of the plot}

\item{.setcolour}{Set the colour of the plot limit lines}

\item{.log.y}{Whether to log-transform the y variable (default = TRUE)}
}
\value{
ggplot2 object with LoD/LoQ evaluation
}
\description{
Create a LoD/LoQ plot
}
