% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual_plots.R
\name{plot_linearity}
\alias{plot_linearity}
\title{Create a linearity plot}
\usage{
plot_linearity(
  data,
  averaged = TRUE,
  .x.offset = 0,
  .log.y = TRUE,
  .log.x = FALSE
)
}
\arguments{
\item{data}{Data to plot}

\item{averaged}{Whether to fit the model for the ungrouped data or not}

\item{.x.offset}{Offset for the linear model on the x-axis (default = 0)}

\item{.log.y}{Whether to log-transform the y variable (default = TRUE)}

\item{.log.x}{Whether to log-transform the x variable (default = FALSE)}
}
\value{
ggplot2 object with linearity evaluation
}
\description{
Create a linearity plot
}
