% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual_plots.R
\name{plot_repeatability}
\alias{plot_repeatability}
\title{Create a repeatability plot}
\usage{
plot_repeatability(data, type = "A", limit = 0.25)
}
\arguments{
\item{data}{Data to plot}

\item{type}{Type of repeatability to plot, "A" = repeatability heatmap, "B" = traditional plot.}

\item{limit}{The acceptance criteria limit}
}
\value{
ggplot2 object with repeatability evaluation
}
\description{
Create a repeatability plot
}
