% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual_plots.R
\name{plot_specificity_recovery}
\alias{plot_specificity_recovery}
\title{Create a specificity recovery plot}
\usage{
plot_specificity_recovery(
  data,
  ref_value = NULL,
  limits = c(0.8, 1.25),
  .log.y = TRUE
)
}
\arguments{
\item{data}{Data to plot}

\item{ref_value}{Reference value to compare against}

\item{limits}{The acceptance criteria limits}

\item{.log.y}{Whether to log-transform the y variable (default = TRUE)}
}
\value{
ggplot2 object with specificity recovery evaluation
}
\description{
Create a specificity recovery plot
}
