% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterPlot.R
\name{rasterPlot}
\alias{rasterPlot}
\title{rasterPlot}
\usage{
rasterPlot(rast, sluCol = "red", classMethod = FALSE, nClass = 5)
}
\arguments{
\item{rast}{The data to be displayed in this layer.}

\item{sluCol}{Choose color scheme. Valid options are "all", "red", "green", "blue", "yellow" or "grey"}

\item{classMethod}{Choose raster classification method. If missing the data is plot as continous. Valid options are "fixed", "sd", "equal", "pretty", "quantile", "kmeans", "hclust", "bclust", "fisher", "jenks", "dpih", "headtails", or "maximum"}

\item{nClass}{number of classes}
}
\value{
A ggplot object
}
\description{
rasterPlot
}
\examples{
f <- system.file("external/test.grd", package="raster")
r <- raster::raster(f)
rasterPlot(rast = r, sluCol = blue, classMethod = "jenks", nClass = 5)
}
