% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rescale}
\alias{rescale}
\title{Re-range some numbers to stretch them so that they fit a scale from a to b}
\usage{
rescale(x, a = 0, b = 1, na.rm = T)
}
\arguments{
\item{x}{A numerical vector}

\item{a}{New minimum for the scale (numeric)}

\item{b}{New maximum for the scale (numeric)}

\item{na.rm}{Whether to remove NAs or not (default = TRUE)}
}
\value{
A stretched version of x that goes from a to b
}
\description{
Re-range some numbers to stretch them so that they fit a scale from a to b
}
\examples{
rescale(1:10, 1, 100)
}
