% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveoutput.R
\name{saveoutput}
\alias{saveoutput}
\title{Save Output}
\usage{
saveoutput(object, filename = NULL, description = NULL)
}
\arguments{
\item{object}{A `trendtest()` object.}

\item{filename}{Your path and file name in a string form. See example below.}

\item{description}{A short description for the analyses.}
}
\description{
This is used to save the output from a `trendtest()` object.
}
\examples{
data <- data.frame(DATE = c("1988-06-27", "1988-08-25", "1990-05-03", "1990-05-31", "1990-06-26", "1990-08-21", "1992-07-16", "1992-08-10", "1994-05-25", "1994-08-03"),
RESPONSE = c(22, 49, 75, 41, 36, 45, 2, 20, 31, 40))
saveoutput(trendtest(data$DATE, data$RESPONSE), file = "path/filename.txt", description = "your description")

}
