% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{scale_colour_catapult_d}
\alias{scale_colour_catapult_d}
\title{Catapult Colour Scales}
\usage{
scale_colour_catapult_d(
  option = "D",
  begin = 0,
  end = 0.85,
  direction = 1,
  legend_size = NULL,
  ...
)
}
\arguments{
\item{option}{The viridis scale to use (A-H)}

\item{begin}{The beginning of the colour scale}

\item{end}{The end of the colour scale}

\item{direction}{The direction of the colour scale}

\item{legend_size}{The size of the legend (discrete)}

\item{...}{Additional arguments to pass to `scale_colour_viridis_d`}
}
\value{
A colour scale to be used in ggplot2
}
\description{
Catapult Colour Scales
}
