% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{scale_fill_catapult_c}
\alias{scale_fill_catapult_c}
\title{Catapult Colour Scales}
\usage{
scale_fill_catapult_c(
  option = "D",
  begin = 0,
  end = 0.85,
  direction = 1,
  legend_width = NULL,
  legend_height = NULL,
  ...
)
}
\arguments{
\item{option}{The viridis scale to use (A-H)}

\item{begin}{The beginning of the colour scale}

\item{end}{The end of the colour scale}

\item{direction}{The direction of the colour scale}

\item{legend_width}{The width of the legend (continuous)}

\item{legend_height}{The height of the legend (continuous)}

\item{...}{Additional arguments to be passed to `scale_fill_viridis_c`}
}
\value{
A colour scale to be used in ggplot2
}
\description{
Catapult Colour Scales
}
