% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{sd.u}
\alias{sd.u}
\title{Unbiased Standard Deviation}
\usage{
sd.u(x)
}
\arguments{
\item{x}{A vector of (normally-distributed) numbers}
}
\value{
The corrected/unbiased standard deviation of the input vector, based on a chi-distributed correction factor (c4)
}
\description{
Unbiased Standard Deviation. Very helpful when sample size is small, especially less than 6,
which is very common in assay qualification or validation runs. Regular sample standard deviation
can be up to 20\% biased when the sample size is 3, for example. This one isn't.
}
\examples{
sd.u(rnorm(10, 0, 4))
}
