% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual_tables.R
\name{table_accuracy}
\alias{table_accuracy}
\title{Table of accuracy}
\usage{
table_accuracy(data, ref_value = NULL, limits = c(0.8, 1.25))
}
\arguments{
\item{data}{Data to calculate accuracy for}

\item{ref_value}{Reference value to compare against. If not provided, relative accuracy compared to group 1 average is used.}

\item{limits}{Limits to compare against as an acceptance criteria}
}
\value{
A table of accuracy
}
\description{
Table of accuracy
}
\examples{
table_accuracy(rnorm(10, 0, 4))
table_accuracy(rnorm(10, 0, 4), 0)
}
