% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual_tables.R
\name{table_comparison}
\alias{table_comparison}
\title{Comparison between methods}
\usage{
table_comparison(x, y)
}
\arguments{
\item{x}{Method 1 data}

\item{y}{Method 2 data}
}
\value{
A table of comparisons between the two methods.
}
\description{
Comparison between methods
}
\examples{
table_comparison(rnorm(10, 0, 4), rnorm(10, 0, 4))
}
