% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual_tables.R
\name{table_intermediate_precision}
\alias{table_intermediate_precision}
\title{Table of intermediate precision}
\usage{
table_intermediate_precision(data, groups = NULL, limit = 0.3)
}
\arguments{
\item{data}{Data to calculate intermediate precision from}

\item{groups}{which groups to include in the intermediate precision assessment}

\item{limit}{The limit to compare against}
}
\value{
A table of intermediate precision
}
\description{
Table of intermediate precision
}
