% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual_tables.R
\name{table_limits}
\alias{table_limits}
\title{Table of Limits}
\usage{
table_limits(data, variance_data = NULL, limit = list(LoD = 3.3, LoQ = 10))
}
\arguments{
\item{data}{Data to calculate baseline from}

\item{variance_data}{Variance data to use. If NULL, the data is used instead}

\item{limit}{The named list of limits to calculate (e.g. 3.3 for LOD, 10 for LOQ)}
}
\value{
A table of limits
}
\description{
Table of Limits
}
\examples{
table_limits(rnorm(10, 0, 4))
table_limits(rnorm(10, 0, 4), rnorm(10, 0, 4))
table_limits(rnorm(10, 0, 4), rnorm(10, 0, 4), list("LoD" = 3.3, "LoQ" = 10))
}
