% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual_tables.R
\name{table_linearity}
\alias{table_linearity}
\title{Table for the Linearity Assessment}
\usage{
table_linearity(data, averaged = TRUE, .log.y = TRUE, .log.x = FALSE)
}
\arguments{
\item{data}{Data to calculate linearity from.}

\item{averaged}{Whether to average the data or not}

\item{.log.y}{Whether to log-transform the y variable}

\item{.log.x}{Whether to log-transform the x variable}
}
\value{
A table of linearity
}
\description{
Table for the Linearity Assessment
}
