% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual_tables.R
\name{table_repeatability}
\alias{table_repeatability}
\title{Table of repeatability}
\usage{
table_repeatability(data, groups = NULL, limit = 0.25)
}
\arguments{
\item{data}{Data to calculate repeatability from}

\item{groups}{which groups to include in the repeatability assessment}

\item{limit}{The limit to compare against}
}
\value{
A table of repeatability
}
\description{
Table of repeatability
}
