% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual_tables.R
\name{table_specificity_recovery}
\alias{table_specificity_recovery}
\title{Table for specificity recovery}
\usage{
table_specificity_recovery(data, ref_value = NULL, limits = c(0.8, 1.25))
}
\arguments{
\item{data}{Data to calculate specificity recovery from}

\item{ref_value}{Reference value to compare against. If not provided, relative accuracy compared to group 1 average is used.}

\item{limits}{Limits to compare against as an acceptance criteria}
}
\value{
A table of specificity recovery
}
\description{
Table for specificity recovery
}
