% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_catapult}
\alias{theme_catapult}
\title{Catapult Theme}
\usage{
theme_catapult(
  ...,
  angle_x = 0,
  angle_y = 0,
  legend.position = "bottom",
  text_size_offset = 0,
  show_gridlines = TRUE
)
}
\arguments{
\item{...}{Additional theme elements}

\item{angle_x}{The angle of the x-axis text}

\item{angle_y}{The angle of the y-axis text}

\item{legend.position}{The position of the legend}

\item{text_size_offset}{The offset for the text size}

\item{show_gridlines}{Whether to show gridlines or not}
}
\value{
A theme to be used in ggplot2
}
\description{
Catapult Theme
}
