% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendtest.dbrda.R
\name{trendtest.dbrda}
\alias{trendtest.dbrda}
\title{Time Series Analyses for Community Data}
\usage{
trendtest.dbrda(x, y, summary = TRUE, nperm = 999, seed = FALSE)
}
\arguments{
\item{x}{A vector of dates or a data frame of the explanatory variables.
If \code{x} is a vector of dates (determined by `is.Date` function of \pkg{lubridate}),
then this function automatically generates a data frame of the DATE, YEAR, MONTH, and YDAY (Julian day).}

\item{y}{Data frame of response variables, where rows represent sites and columns represent species.}

\item{summary}{\code{TRUE} returns a summarized version of the db-RDA results. \code{FALSE} returns the full uncondensed results in a \code{list} object.}

\item{nperm}{Number of permutations. \code{999} is default.}

\item{seed}{\code{FALSE} is default. Otherwise, it can be any \code{numeric} to set up the seed using `set.seed(seed)`, which ensure the results from permutation steps are replicable.}
}
\description{
This function tests for the interannual and seasonal shifts in a given community composition.
It is based on Bray-Curtis dissimilarity distance-based redundancy analysis. \cr\cr

This is primarily a wrapper for the following functions in \pkg{vegan} and various functions in \pkg{tidyverse} for data management. Please cite the listed packages.
\itemize{
\item \pkg{vegan}:
\code{`decostand`} for Hellinger standardization,
\code{`vegdist`} for calculating Bray-Curtis dissimilarity matrix,
\code{`adonis2`} for perMANOVA,
\code{`dbrda`} for distanced-based RDA,
\code{`sppscores`} for generating species scores,
\code{`score`} for extracting scores,
\code{`ordiR2step`} for forward selection,
\code{`varpart`} for variance partitioning.
}
}
\examples{
data <- data.frame(DATE = c("1988-06-27", "1988-08-25", "1990-05-03", "1990-05-31", "1990-06-26", "1990-08-21", "1992-07-16", "1992-08-10", "1994-05-25", "1994-08-03"),
Sp1 = c(22, 49, 75, 41, 36, 45, 2, 20, 31, 40),
Sp2 = c(70, 58, 96, 28, 78, 79, 53, 66, 84, 80))
trendtest.dbrda(data$DATE, data[-1])

}
