% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendtest.R
\name{trendtest}
\alias{trendtest}
\title{Time Series Analyses for Water Quality Data}
\usage{
trendtest(x, y, summary = TRUE, median = FALSE)
}
\arguments{
\item{x}{Vector of the dates corresponding to \code{y}.}

\item{y}{Vector of the response variable.}

\item{summary}{If \code{TRUE (DEFAULT)}, it returns the summary output for the results.
If \code{FALSE}, it returns the full native results in a \code{list} object}

\item{median}{If \code{TRUE}, the Mann-Kendall test is performed on annual median.
If \code{FALSE (DEFAULT)}, it is performed on all values while correcting for ties in the estimation.}
}
\description{
The set of analyses in this are adapted from Alberta Lake Management Society (ALMS) 2018: Guide to trend analysis on Alberta lakes. \cr
(\url{https://alms.ca/wp-content/uploads/2018/04/ALMS-Trend-Analysis-Methods-04-2018.pdf}) \cr\cr

This is primarily a wrapper for the functions in the following packages, including various functions in \pkg{tidyverse} for data management. Please cite the listed packages.
\itemize{
\item \pkg{FSA}: \code{`dunnTest`}
\item \pkg{EnvStats}: \code{`kendallSeasonalTrendTest`} and \code{`kendallTrendTest`}
\item \pkg{car}: \code{`leveneTest`}
}
}
\examples{
data <- data.frame(DATE = c("1988-06-27", "1988-08-25", "1990-05-03", "1990-05-31", "1990-06-26", "1990-08-21", "1992-07-16", "1992-08-10", "1994-05-25", "1994-08-03"),
RESPONSE = c(22, 49, 75, 41, 36, 45, 2, 20, 31, 40))
trendtest(data$DATE, data$RESPONSE)

}
