% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{logCV}
\alias{logCV}
\title{Coefficient of Variation for Log-transformed Data}
\usage{
logCV(x, na.rm = F)
}
\arguments{
\item{x}{A vector of log-normally distributed data}

\item{na.rm}{Whether or not to ignore NAs}
}
\description{
A less known fact is that for log-transformed data you can't just take 100% * σ / μ and call it a day.
This function implements the correct way of getting a CV when you apply it on log-transformed data.
It does a method of moments based estimation of it, to check full maths and everything, read the paper published by Nelson W. in Applied Life Data Analysis. USA: John Wiley & Sons Inc; 2003.
}
\examples{
logCV(c(0.8, 0.9, 0.78))
}
