% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_in_excel.R
\name{open_in_excel}
\alias{open_in_excel}
\title{Open in Excel}
\arguments{
\item{.data}{A data.frame that you want to open in excel}
}
\value{
Returns same data as input
}
\description{
Open in Excel
}
\examples{
open_in_excel(iris)
}
