% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{pvl}
\alias{pvl}
\title{Get pretty p-values (sometimes)}
\usage{
pvl(x, signif = 4)
}
\arguments{
\item{x}{A vector of p-values}

\item{signif}{Significant digits (default = 4) to keep}
}
\value{
x, if x > x^(-signif), otherwise "<1^(-signif)" in non-scientific notation
}
\description{
Get pretty p-values (sometimes)
}
\examples{
pvl(c(0.0000003, 0.06), 2) # returns "<0.01" "0.06"
}
